Moved PasswordAlgorithm to restcomm.commons and added 'password algorithm strategy' configuration option (in java)
Refers #1497
