+++ restcomm/restcomm.http/src/test/java/org/restcomm/connect/http/EndpointMockedTest.java
@@ -23,6 +23,7 @@ package org.restcomm.connect.http;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.XMLConfiguration;
+import org.restcomm.connect.commons.security.PasswordAlgorithm;
 import org.restcomm.connect.dao.AccountsDao;
 import org.restcomm.connect.dao.AccountsDaoMock;
 import org.restcomm.connect.dao.DaoManager;
@@ -73,7 +74,7 @@ public class EndpointMockedTest {
         when(servletContext.getAttribute(Configuration.class.getName())).thenReturn(conf);
         // mock accountsDao
         accounts = new ArrayList<Account>();
-        accounts.add(new Account(new Sid("AC00000000000000000000000000000000"),null,null,"administrator@company.com","Administrator",null,null,null, "RestComm", Account.PasswordAlgorithm.plain, "77f8c12cc7b8f8423e5c38b035249166",null,null));
+        accounts.add(new Account(new Sid("AC00000000000000000000000000000000"),null,null,"administrator@company.com","Administrator",null,null,null, "RestComm", PasswordAlgorithm.plain, "77f8c12cc7b8f8423e5c38b035249166",null,null));
         accountsDao = new AccountsDaoMock(accounts);
         // mock DaoManager
         daoManager = new DaoManagerMock(accountsDao);
